refactor: move tx to the ids linked to the executed nonces
