+++ multisig/simulations/ethereum/vip-045/index.ts
@@ -1,73 +1,105 @@
 import { expect } from "chai";
 import { Contract } from "ethers";
 import { ethers } from "hardhat";
-import { NETWORK_ADDRESSES } from "src/networkAddresses";
-import { forking, pretendExecutingVip } from "src/vip-framework";
 
+import { forking, pretendExecutingVip } from "../../../../src/vip-framework";
 import vip045, {
-  ETHEREUM_ACM,
-  ETHEREUM_NORMAL_TIMELOCK,
-  ETHEREUM_OMNICHAIN_EXECUTOR_OWNER,
+  PRIME,
+  PRIME_LIQUIDITY_PROVIDER,
+  USDC,
+  USDC_PER_90_DAYS_REWARD,
+  USDC_PER_BLOCK_REWARD,
+  USDT,
+  USDT_PER_90_DAYS_REWARD,
+  USDT_PER_BLOCK_REWARD,
+  WBTC,
+  WBTC_PER_90_DAYS_REWARD,
+  WBTC_PER_BLOCK_REWARD,
+  WETH,
+  WETH_PER_90_DAYS_REWARD,
+  WETH_PER_BLOCK_REWARD,
 } from "../../../proposals/ethereum/vip-045";
-import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManagerAbi.json";
+import ERC20_ABI from "./abis/ERC20.json";
+import PRIME_ABI from "./abis/Prime.json";
+import PRIME_LIQUIDITY_PROVIDER_ABI from "./abis/PrimeLiquidityProvider.json";
 
-const { ethereum } = NETWORK_ADDRESSES;
+forking(20311017, async () => {
+  const erc20At = (address: string) => new ethers.Contract(address, ERC20_ABI, ethers.provider);
 
-forking(20039640, async () => {
-  let acm: Contract;
-  let defaultAdminRole: string;
-  before(async () => {
-    acm = await ethers.getContractAt(ACCESS_CONTROL_MANAGER_ABI, ETHEREUM_ACM);
-    defaultAdminRole = await acm.DEFAULT_ADMIN_ROLE();
-  });
-  describe("Pre-VIP behaviour", async () => {
-    it("Normal Timelock does not has default admin role", async () => {
-      const hasRole = await acm.hasRole(defaultAdminRole, ETHEREUM_NORMAL_TIMELOCK);
-      expect(hasRole).equals(false);
+  describe("Pre-VIP behavior", () => {
+    let primeLiquidityProvider: Contract;
+
+    before(async () => {
+      primeLiquidityProvider = await ethers.getContractAt(PRIME_LIQUIDITY_PROVIDER_ABI, PRIME_LIQUIDITY_PROVIDER);
     });
 
-    it("Guardian is not allowed to call retryMessage", async () => {
-      const role = ethers.utils.solidityPack(
-        ["address", "string"],
-        [ETHEREUM_OMNICHAIN_EXECUTOR_OWNER, "retryMessage(uint16,bytes,uint64,bytes)"],
-      );
-      const roleHash = ethers.utils.keccak256(role);
-      expect(await acm.hasRole(roleHash, ethereum.GUARDIAN)).to.be.false;
+    it("speeds", async () => {
+      let speed = await primeLiquidityProvider.tokenDistributionSpeeds(WETH);
+      expect(speed).to.deep.equal(0);
+
+      speed = await primeLiquidityProvider.tokenDistributionSpeeds(WBTC);
+      expect(speed).to.deep.equal(0);
+
+      speed = await primeLiquidityProvider.tokenDistributionSpeeds(USDC);
+      expect(speed).to.deep.equal(0);
+
+      speed = await primeLiquidityProvider.tokenDistributionSpeeds(USDT);
+      expect(speed).to.deep.equal(0);
     });
 
-    it("Guardian is not allowed to call forceResumeReceive", async () => {
-      const role = ethers.utils.solidityPack(
-        ["address", "string"],
-        [ETHEREUM_OMNICHAIN_EXECUTOR_OWNER, "forceResumeReceive(uint16,bytes)"],
-      );
-      const roleHash = ethers.utils.keccak256(role);
-      expect(await acm.hasRole(roleHash, ethereum.GUARDIAN)).to.be.false;
+    it("paused", async () => {
+      const paused = await primeLiquidityProvider.paused();
+      expect(paused).to.be.equal(true);
     });
   });
+
   describe("Post-VIP behavior", async () => {
+    let primeLiquidityProvider: Contract;
+    let prime: Contract;
+
     before(async () => {
       await pretendExecutingVip(await vip045());
-    });
-    it("Normal Timelock has default admin role", async () => {
-      const hasRole = await acm.hasRole(defaultAdminRole, ETHEREUM_NORMAL_TIMELOCK);
-      expect(hasRole).equals(true);
-    });
-    it("Guardian is allowed to call retryMessage", async () => {
-      const role = ethers.utils.solidityPack(
-        ["address", "string"],
-        [ETHEREUM_OMNICHAIN_EXECUTOR_OWNER, "retryMessage(uint16,bytes,uint64,bytes)"],
-      );
-      const roleHash = ethers.utils.keccak256(role);
-      expect(await acm.hasRole(roleHash, ethereum.GUARDIAN)).to.be.true;
+
+      primeLiquidityProvider = await ethers.getContractAt(PRIME_LIQUIDITY_PROVIDER_ABI, PRIME_LIQUIDITY_PROVIDER);
+      prime = await ethers.getContractAt(PRIME_ABI, PRIME);
     });
 
-    it("Guardian is allowed to call forceResumeReceive", async () => {
-      const role = ethers.utils.solidityPack(
-        ["address", "string"],
-        [ETHEREUM_OMNICHAIN_EXECUTOR_OWNER, "forceResumeReceive(uint16,bytes)"],
-      );
-      const roleHash = ethers.utils.keccak256(role);
-      expect(await acm.hasRole(roleHash, ethereum.GUARDIAN)).to.be.true;
+    it("speeds", async () => {
+      let speed = await primeLiquidityProvider.tokenDistributionSpeeds(WETH);
+      expect(speed).to.deep.equal(WETH_PER_BLOCK_REWARD);
+
+      speed = await primeLiquidityProvider.tokenDistributionSpeeds(WBTC);
+      expect(speed).to.deep.equal(WBTC_PER_BLOCK_REWARD);
+
+      speed = await primeLiquidityProvider.tokenDistributionSpeeds(USDC);
+      expect(speed).to.deep.equal(USDC_PER_BLOCK_REWARD);
+
+      speed = await primeLiquidityProvider.tokenDistributionSpeeds(USDT);
+      expect(speed).to.deep.equal(USDT_PER_BLOCK_REWARD);
+    });
+
+    it("paused", async () => {
+      const paused = await primeLiquidityProvider.paused();
+      expect(paused).to.be.equal(false);
+
+      const primePaused = await prime.paused();
+      expect(primePaused).to.be.equal(false);
+    });
+
+    it("has enough WBTC for the reward", async () => {
+      expect(await erc20At(WBTC).balanceOf(PRIME_LIQUIDITY_PROVIDER)).to.be.greaterThanOrEqual(WBTC_PER_90_DAYS_REWARD);
+    });
+
+    it("has enough WETH for the reward", async () => {
+      expect(await erc20At(WETH).balanceOf(PRIME_LIQUIDITY_PROVIDER)).to.be.greaterThanOrEqual(WETH_PER_90_DAYS_REWARD);
+    });
+
+    it("has enough USDC for the reward", async () => {
+      expect(await erc20At(USDC).balanceOf(PRIME_LIQUIDITY_PROVIDER)).to.be.greaterThanOrEqual(USDC_PER_90_DAYS_REWARD);
+    });
+
+    it("has enough USDT for the reward", async () => {
+      expect(await erc20At(USDT).balanceOf(PRIME_LIQUIDITY_PROVIDER)).to.be.greaterThanOrEqual(USDT_PER_90_DAYS_REWARD);
     });
   });
 });
